Move etl.timer to etl.util.timer
